pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dora"
version = "0.1.0"
description = "Multi-Agent Event Notification System - A distributed system with Local Calendar Intelligence, Message Generation, Audience Analysis, and Image Generation agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    # Core framework
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    # Agent framework
    "openai-agents>=0.0.14",
    # A2A Protocol dependencies
    "pydantic-ai[a2a]>=0.0.9",
    # Database and caching
    "redis>=4.5.0",
    "psycopg2-binary>=2.9.6",
    "asyncpg>=0.28.0",
    # AI/ML providers
    "openai>=1.1.0",
    "anthropic>=0.7.0",
    # MCP tools
    "mcp",
    # Testing
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    # Utilities
    "nest-asyncio>=1.5.0",
    "aiohttp>=3.12.2",
    "python-multipart>=0.0.6",
    "structlog>=23.1.0",
    "jsonschema>=4.24.0",
    "a2a-sdk>=0.2.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
    "coverage>=7.0.0",
    "pytest-cov>=4.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "B", "C4", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
strict = true

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["dora"]
